fix(demo): issue with creating multiple suggestions for the same match
