+++ apps/website/dom.d.ts
@@ -1,0 +1,3 @@
+interface ObservableHTMLElement extends HTMLElement {
+  _observer?: MutationObserver;
+}
