+++ apps/website/hooks/useDomLayoutChange.ts
@@ -1,0 +1,13 @@
+import { useEffect } from 'react';
+
+export const useDomLayoutChange = (callback: () => void) => {
+  useEffect(() => {
+    window.addEventListener('scroll', callback, true);
+    window.addEventListener('resize', callback, true);
+
+    return () => {
+      window.removeEventListener('scroll', callback, true);
+      window.removeEventListener('resize', callback, true);
+    };
+  }, []);
+};
