+++ apps/website/package.json
@@ -12,12 +12,14 @@
   "dependencies": {
     "@floating-ui/dom": "^1.6.11",
     "@floating-ui/react": "^0.26.24",
+    "@lit/react": "^1.0.6",
     "@radix-ui/react-icons": "^1.3.0",
     "@radix-ui/react-popover": "^1.1.1",
     "@radix-ui/react-slot": "^1.1.0",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
     "embla-carousel-react": "^8.3.0",
+    "lit": "^3.2.1",
     "lucide-react": "^0.439.0",
     "next": "14.2.7",
     "react": "^18",
